+++ packages/features/src/pots/PotFormContent/index.ts
@@ -1,0 +1,6 @@
+export { PotFormContent } from './PotFormContent.native'
+export type {
+  IPotFormContentProps,
+  IPotFormRef,
+  IPotFormValues,
+} from './PotFormContent'