+++ packages/features/src/pots/index.web.ts
@@ -1,2 +1,20 @@
 export { PotCard } from './PotCard/index.web'
 export type { IPotCardProps } from './PotCard/index.web'
+
+export { PotFormContent } from './PotFormContent/index.web'
+export type {
+  IPotFormContentProps,
+  IPotFormRef,
+  IPotFormValues,
+} from './PotFormContent/index.web'
+
+export {
+  createAddPotModalConfig,
+  createEditPotModalConfig,
+  createDeletePotModalConfig,
+} from './createPotModalConfigs'
+export type {
+  IAddPotModalLabels,
+  IEditPotModalLabels,
+  IDeletePotModalLabels,
+} from './createPotModalConfigs'